+++ CHANGES.md
@@ -45,6 +45,11 @@
 - Add more detailed pragma registration.
 - Add checks MissingPragma, PragmaInvalidClassifyLevelCheck, PragmaInvalidUsageCheck.
 - Add f4 t keybinding to VSCode plugin to write a trace command.
+- Fix determining assigned variable in top scope being a global.
+- Fix `ScopeEntry` types when ending up in the global scope.
+- Fix `HidesVariableCheck` not seeing `DEFINITION` `ScopeEntry`s.
+- Add support for inline `# mlint: disable=all` to disable all checks.
+- Fix `VariableDeclarationUsageDistanceCheck` to see augmented assignment (e.g. `_andif<<` or `+<< 1`) as valid usage.
 - Several fixes.
 
 ### Breaking changes (reiterated from above)